fix opacity type
darcs-hash:20070921225522-2fc9d-e8d90dd1098885248ce893ec2334bcf1788e4a96.gz

+++ label_overlay.cxx
@@ -39,7 +39,7 @@ int main(int argc, char * argv[])
   ColorizerType::Pointer colorizer = ColorizerType::New();
   colorizer->SetInput( converter->GetOutput() );
   colorizer->SetFeatureImage( reader2->GetOutput() );
-  colorizer->SetOpacity( atoi(argv[6]) );
+  colorizer->SetOpacity( atof(argv[6]) );
 
   itk::SimpleFilterWatcher watcher(colorizer, "filter");
